requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
tomli==2.2.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
wsproto==1.2.0
